web: refactor 2/7 federation
